development: get Combo fields working
Map COMBO fields to VirtualFields. Virtual fields will initially be a
superset of COMBO and SYNTH fileds.  The only difference between those 2
field types is that the field separators are optional for COMBO fields
and required for SYNTH fields.
Virtual fields will eventually have the ability to include arbitrary
text.  Think of them as fmt fields.